requirements.txt
# Lightweight dependencies for deployment
Flask
flask-cors
python-dotenv
pymongo
requests
openai
serverless-wsgi
werkzeug
gunicorn

# Heavy ML libraries removed for Vercel/Lambda constraints:
# language-tool-python
# textstat
# langchain
# langchain-community
# langchain-chroma
# langchain-text-splitters
# sentence-transformers
# pypdf
# tiktoken